requirements.txt
scapy>=2.5.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
requests>=2.31.0
pyyaml>=6.0.1
click>=8.1.7
rich>=13.0.0
# Authentication & security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# Rate limiting
slowapi>=0.1.9
# Dev / testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
ruff>=0.3.0
black>=24.0.0
isort>=5.13.0
bandit>=1.7.7
pip-audit>=2.7.0
